Rename anakinparse in logic
