+++ src/main/java/seedu/address/logic/anakinparser/AnakinEditDeckCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.anakinParser;
+package seedu.address.logic.anakinparser;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
