+++ src/main/java/seedu/address/logic/anakinparser/AnakinParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.anakinParser;
+package seedu.address.logic.anakinparser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
